refactor: name the saved-report tools after the UI's "Custom Reports"
Actual's UI calls these Custom Reports — they appear as a category inside the
"Add new widget" menu — while "widget" refers to the dashboard card itself.
Naming them get-reports/create-report also collided with the analysis report
tools, so rename to *-custom-report and move them under src/tools/custom-reports.

The sidebar labels the section "Reports" while the pages within it are called
dashboards, so the dashboard tool descriptions now mention the Reports section
to keep them findable under either term.
