+++ src/tools/custom-reports/input-schema.ts
